Bookworm rootfs 

Commits collected and merged

+++ recipes/devices/pi.sh
@@ -33,32 +33,26 @@ KIOSKMODE=no
 ## Partition info
 BOOT_START=0
 BOOT_END=96
-BOOT_TYPE=msdos    # msdos or gpt
-BOOT_USE_UUID=yes  # Add UUID to fstab
+BOOT_TYPE=msdos   # msdos or gpt
+BOOT_USE_UUID=yes # Add UUID to fstab
 INIT_TYPE="initv3"
 
 ## Plymouth theme management
-PLYMOUTH_THEME="volumio-player"	# Choices are: {volumio,volumio-logo,volumio-player}
-INIT_PLYMOUTH_DISABLE="no"		# yes/no or empty. Removes plymouth initialization in init if "yes" is selected
+PLYMOUTH_THEME="volumio-player" # Choices are: {volumio,volumio-logo,volumio-player}
+INIT_PLYMOUTH_DISABLE="no"      # yes/no or empty. Removes plymouth initialization in init if "yes" is selected
 
 ## TODO: for any KMS DRM panel mudule, which does not create frambuffer bridge, set this variable to yes, otherwise no
 ## Implement an if/else statement to handle this properly
-UPDATE_PLYMOUTH_SERVICES_FOR_KMS_DRM="no"	# yes/no or empty. Replaces default plymouth systemd services if "yes" is selected
+UPDATE_PLYMOUTH_SERVICES_FOR_KMS_DRM="no" # yes/no or empty. Replaces default plymouth systemd services if "yes" is selected
 
 # Modules that will be added to initramfs
 MODULES=("drm" "fuse" "nls_cp437" "nls_iso8859_1" "nvme" "nvme_core" "overlay" "squashfs" "uas")
 # Packages that will be installed
-PACKAGES=( # Bluetooth packages
-	"bluez" "bluez-firmware" "pi-bluetooth"
-	# Foundation stuff
-	"raspberrypi-sys-mods"
+PACKAGES=(
 	# Framebuffer stuff
-	"fbset"	
-	# "rpi-eeprom"\ Needs raspberrypi-bootloader that we hold back
+	"fbset"
 	# GPIO stuff
-	"wiringpi"	
-	# Wireless firmware
-	"firmware-atheros" "firmware-ralink" "firmware-realtek" "firmware-brcm80211"
+	# "wiringpi"
 	# Install to disk tools
 	"liblzo2-2" "squashfs-tools"
 )
@@ -101,10 +95,6 @@ write_device_bootloader() {
 
 # Will be called by the image builder for any customisation
 device_image_tweaks() {
-	# log "Custom dtoverlay pre and post" "ext"
-	# mkdir -p "${ROOTFSMNT}/opt/vc/bin/"
-	# cp -rp "${SRC}"/volumio/opt/vc/bin/* "${ROOTFSMNT}/opt/vc/bin/"
-
 	log "Fixing hostapd.conf" "info"
 	cat <<-EOF >"${ROOTFSMNT}/etc/hostapd/hostapd.conf"
 		interface=wlan0
@@ -126,9 +116,9 @@ device_image_tweaks() {
 
 	log "Adding archive.raspberrypi debian repo" "info"
 	cat <<-EOF >"${ROOTFSMNT}/etc/apt/sources.list.d/raspi.list"
-		deb http://archive.raspberrypi.org/debian/ buster main ui
+		deb http://archive.raspberrypi.org/debian/ ${SUITE} main
 		# Uncomment line below then 'apt-get update' to enable 'apt-get source'
-		#deb-src http://archive.raspberrypi.org/debian/ buster main ui
+		#deb-src http://archive.raspberrypi.org/debian/ ${SUITE} main
 	EOF
 
 	# raspberrypi-{kernel,bootloader} packages update kernel & firmware files
@@ -159,8 +149,8 @@ device_image_tweaks() {
 	# BOOT_PATH=${ROOT_PATH}/boot
 
 	log "Copying custom initramfs script functions" "cfg"
-	[ -d ${ROOTFSMNT}/root/scripts ] || mkdir ${ROOTFSMNT}/root/scripts
-	cp "${SRC}/scripts/initramfs/custom/pi/custom-functions" ${ROOTFSMNT}/root/scripts
+	[ -d "${ROOTFSMNT}"/root/scripts ] || mkdir "${ROOTFSMNT}"/root/scripts
+	cp "${SRC}/scripts/initramfs/custom/pi/custom-functions" "${ROOTFSMNT}"/root/scripts
 }
 
 # Will be run in chroot (before other things)
@@ -228,12 +218,12 @@ device_chroot_tweaks_pre() {
 	# List of custom firmware -
 	# github archives that can be extracted directly
 	declare -A CustomFirmware=(
-		[AlloPiano]="https://github.com/allocom/piano-firmware/archive/master.tar.gz"
-		[TauDAC]="https://github.com/taudac/modules/archive/rpi-volumio-${KERNEL_VERSION}-taudac-modules.tar.gz"
-		[Bassowl]="https://raw.githubusercontent.com/Darmur/bassowl-hat/master/driver/archives/modules-rpi-${KERNEL_VERSION}-bassowl.tar.gz"
-		[wm8960]="https://raw.githubusercontent.com/hftsai256/wm8960-rpi-modules/main/wm8960-modules-rpi-${KERNEL_VERSION}.tar.gz"
-		[brcmfmac43430b0]="https://raw.githubusercontent.com/volumio/volumio3-os-static-assets/master/firmwares/brcmfmac43430b0/brcmfmac43430b0.tar.gz"
-		[PiCustom]="https://raw.githubusercontent.com/Darmur/volumio-rpi-custom/main/output/modules-rpi-${KERNEL_VERSION}-custom.tar.gz"
+		# [AlloPiano]="https://github.com/allocom/piano-firmware/archive/master.tar.gz"
+		# [TauDAC]="https://github.com/taudac/modules/archive/rpi-volumio-${KERNEL_VERSION}-taudac-modules.tar.gz"
+		# [Bassowl]="https://raw.githubusercontent.com/Darmur/bassowl-hat/master/driver/archives/modules-rpi-${KERNEL_VERSION}-bassowl.tar.gz"
+		# [wm8960]="https://raw.githubusercontent.com/hftsai256/wm8960-rpi-modules/main/wm8960-modules-rpi-${KERNEL_VERSION}.tar.gz"
+		# [brcmfmac43430b0]="https://raw.githubusercontent.com/volumio/volumio3-os-static-assets/master/firmwares/brcmfmac43430b0/brcmfmac43430b0.tar.gz"
+		# [PiCustom]="https://raw.githubusercontent.com/Darmur/volumio-rpi-custom/main/output/modules-rpi-${KERNEL_VERSION}-custom.tar.gz"
 	)
 
 	### Kernel installation
@@ -244,17 +234,6 @@ device_chroot_tweaks_pre() {
 	log "Adding kernel ${KERNEL_VERSION} using rpi-update" "info"
 	log "Fetching SHA: ${KERNEL_COMMIT} from branch: ${KERNEL_BRANCH}" "info"
 	echo y | SKIP_BACKUP=1 WANT_32BIT=1 WANT_64BIT=1 WANT_PI4=1 WANT_PI5=0 SKIP_CHECK_PARTITION=1 UPDATE_SELF=0 BRANCH=${KERNEL_BRANCH} /usr/bin/rpi-update "${KERNEL_COMMIT}"
-
-	log "Adding Custom firmware from github" "info"
-	for key in "${!CustomFirmware[@]}"; do
-		wget -nv "${CustomFirmware[$key]}" -O "$key.tar.gz" || {
-			log "Failed to get firmware:" "err" "${key}"
-			rm "$key.tar.gz"
-			continue
-		}
-		tar --strip-components 1 --exclude "*.hash" --exclude "*.md" -xf "$key.tar.gz"
-		rm "$key.tar.gz"
-	done
 
 	## Comment to keep RPi4/RPi5 64bit kernel
 	#if [ -d "/lib/modules/${KERNEL_VERSION}-v8+" ]; then
@@ -278,13 +257,19 @@ device_chroot_tweaks_pre() {
 	log "Finished Kernel installation" "okay"
 
 	### Other Rpi specific stuff
-	log "Installing fake libraspberrypi0 package" "info"
-	wget -nv https://github.com/volumio/volumio3-os-static-assets/raw/master/custom-packages/libraspberrypi0/libraspberrypi0_1.20230509-buster-1_armhf.deb
-	dpkg -i libraspberrypi0_1.20230509-buster-1_armhf.deb
-	rm libraspberrypi0_1.20230509-buster-1_armhf.deb
-
+	
 	## Lets update some packages from raspbian repos now
 	apt-get update && apt-get -y upgrade
+
+	# https://github.com/volumio/volumio3-os/issues/174
+	## Quick fix for dhcpcd in Raspbian vs Debian
+	log "Raspbian vs Debian dhcpcd debug "
+	apt-get remove dhcpcd -yy && apt-get autoremove
+	# wget -nv http://ftp.debian.org/debian/pool/main/d/dhcpcd5/dhcpcd-base_9.4.1-24~deb12u4_armhf.deb
+	# wget -nv http://ftp.debian.org/debian/pool/main/d/dhcpcd5/dhcpcd_9.4.1-24~deb12u4_all.deb
+	wget -nv https://github.com/volumio/volumio3-os-static-assets/raw/master/custom-packages/dhcpcd/dhcpcd_9.4.1-24~deb12u4_all.deb
+	wget -nv https://github.com/volumio/volumio3-os-static-assets/raw/master/custom-packages/dhcpcd/dhcpcd-base_9.4.1-24~deb12u4_armhf.deb
+	dpkg -i dhcpcd*.deb
 
 	NODE_VERSION=$(node --version)
 	log "Node version installed:" "dbg" "${NODE_VERSION}"
@@ -348,19 +333,45 @@ device_chroot_tweaks_pre() {
 	log "Updating LD_LIBRARY_PATH" "info"
 	ldconfig
 
+	# Fixing some Debian/sid quirks
+	if dpkg -l libraspberrypi0 &>/dev/null && [[ "$(dpkg --print-architecture)" = armhf ]]; then
+		log "Fixing linker ld-linux-armhf"
+		ln -s /lib/ld-linux-armhf.so.3 /lib/ld-linux.so.3 2>/dev/null || true
+		ln -s /lib/arm-linux-gnueabihf/ld-linux-armhf.so.3 /lib/arm-linux-gnueabihf/ld-linux.so.3 2>/dev/null || true
+	fi
+	
 	log "Symlinking vc bins" "info"
+	# Clean this up! > Quoting popcornmix "Code from here is no longer installed on latest RPiOS Bookworm images.If you are using code from here you should rethink your solution.Consider this repo closed."
 	# https://github.com/RPi-Distro/firmware/blob/debian/debian/libraspberrypi-bin.links
 	VC_BINS=("edidparser" "raspistill" "raspivid" "raspividyuv" "raspiyuv"
-		"tvservice" "vcdbg" "vcgencmd" "vchiq_test"
-		"dtoverlay" "dtoverlay-pre" "dtoverlay-post" "dtmerge")
+		"tvservice" "vcdbg" "vchiq_test"
+		"dtoverlay-pre" "dtoverlay-post")
+
 	for bin in "${VC_BINS[@]}"; do
-		ln -s "/opt/vc/bin/${bin}" "/usr/bin/${bin}"
+		if [[ ! -f /usr/bin/${bin} && -f /opt/vc/bin/${bin} ]]; then
+			ln -s "/opt/vc/bin/${bin}" "/usr/bin/${bin}"
+			log "Linked ${bin}"
+		else
+			log "${bin} wasn't linked!" "wrn"
+		fi
 	done
 
 	log "Fixing vcgencmd permissions" "info"
 	cat <<-EOF >/etc/udev/rules.d/10-vchiq.rules
 		SUBSYSTEM=="vchiq",GROUP="video",MODE="0660"
 	EOF
+
+	# All additional drivers
+	log "Adding Custom firmware from github" "info"
+	for key in "${!CustomFirmware[@]}"; do
+		wget -nv "${CustomFirmware[$key]}" -O "$key.tar.gz" || {
+			log "Failed to get firmware:" "err" "${key}"
+			rm "$key.tar.gz"
+			continue
+		}
+		tar --strip-components 1 --exclude "*.hash" --exclude "*.md" -xf "$key.tar.gz"
+		rm "$key.tar.gz"
+	done
 
 	# Rename gpiomem in udev rules if kernel is equal or greater than 6.1.54
 	if [ "$MAJOR_VERSION" -gt 6 ] || { [ "$MAJOR_VERSION" -eq 6 ] && { [ "$MINOR_VERSION" -gt 1 ] || [ "$MINOR_VERSION" -eq 1 ] && [ "$PATCH_VERSION" -ge 54 ]; }; }; then
@@ -511,9 +522,9 @@ device_chroot_tweaks_post() {
 # Will be called by the image builder post the chroot, before finalisation
 device_image_tweaks_post() {
 	log "Running device_image_tweaks_post" "ext"
-    # Plymouth systemd services OVERWRITE
+	# Plymouth systemd services OVERWRITE
 	if [[ "${UPDATE_PLYMOUTH_SERVICES}" == yes ]]; then
-        log "Updating plymouth systemd services" "info"
-        cp -dR "${SRC}"/volumio/framebuffer/systemd/* "${ROOTFSMNT}"/lib/systemd
+		log "Updating plymouth systemd services" "info"
+		cp -dR "${SRC}"/volumio/framebuffer/systemd/* "${ROOTFSMNT}"/lib/systemd
 	fi
 }
